requirements.txt
# requirements.txt

fastapi
uvicorn
python-multipart

# For audio processing and ML models
torch
torchaudio
numpy

# Direct FFmpeg bindings for robust decoding
ffmpeg-python

# Source Separation
demucs

# Automatic Speech Recognition
openai-whisper